Added validation to createInidividualCustomer and an alert screen
Once the user tries to create an account, the name entered is
checked as to whether it is already in the database. If so, an
alert screen pops up notifying the user that the name already has
a registered account and that they should create an account under
an unused name. If the name is not already registered, the info
is stored and a new account is created.